move `uuid` to directory `src`.
