+++ src/index.js
@@ -7,4 +7,4 @@ export {default as convertFromJSON} from './convertFromJSON';
 export {default as convertToJSON} from './convertToJSON';
 export {default as createEmptyEditorState} from './createEmptyEditorState';
 export {default as isEditorStateEmpty} from './isEditorStateEmpty';
-export {default as uuid} from './ui/uuid';
+export {default as uuid} from './uuid';
